Improve UIH detection
Transactions with multiple inputs of the same previous output's script are possibly
due to the preference to always spend all inputs of the same script together, which
improves privacy and is implemented by (at least) Bitcoin Core and Electrum.

This behaviour looks like UIH, but should not trigger our UIH detection.
We now opt these transactions out of the UIH tests.